refactor: 페이지 관련 응답 DTO 패키지 구조 변경
