+++ src/main/java/com/greedy/mokkoji/api/recruitment/service/RecruitmentService.java
@@ -1,6 +1,6 @@
 package com.greedy.mokkoji.api.recruitment.service;
 
-import com.greedy.mokkoji.api.club.dto.page.PageResponse;
+import com.greedy.mokkoji.api.pagination.dto.PageResponse;
 import com.greedy.mokkoji.api.external.AppDataS3Client;
 import com.greedy.mokkoji.api.recruitment.dto.response.AllRecruitment.AllRecruitmentResponse;
 import com.greedy.mokkoji.api.recruitment.dto.response.AllRecruitment.ClubPreviewResponse;
